refactor: Move question entity into entities folder
