+++ lib/external/datasources/question.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 import 'package:http/http.dart' as http;
-import 'package:pub_puzzler/domain/question.dart';
+import 'package:pub_puzzler/domain/entities/question.dart';
 import 'package:pub_puzzler/domain/question_repository.dart';
 
 const String apiURL = 'https://opentdb.com/api.php';
