+++ lib/presenter/question_card.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:pub_puzzler/domain/question.dart';
+import 'package:pub_puzzler/domain/entities/question.dart';
 import 'package:pub_puzzler/infra/services/game_provider.dart';
 import 'package:pub_puzzler/external/datasources/question.dart';
 import 'package:pub_puzzler/presenter/answer_tile.dart';
